Add dummy tests for the constructors, getters and setters. Refactor by introducing FallibleEntity
