--- app/src/main/java/ch/epfl/qedit/Request.java
@@ -1,13 +1,0 @@
-package ch.epfl.qedit;
-
-public abstract class Request {
-    private Status status;
-
-    public Status getStatus() {
-        return status;
-    }
-
-    public void setStatus(Status status) {
-        this.status = status;
-    }
-}
