--- app/src/main/java/ch/epfl/qedit/Response.java
@@ -1,13 +1,0 @@
-package ch.epfl.qedit;
-
-public abstract class Response {
-    private Status status;
-
-    public Status getStatus() {
-        return status;
-    }
-
-    protected void setStatus(Status status) {
-        this.status = status;
-    }
-}
