--- app/src/main/java/ch/epfl/qedit/ResponseLogin.java
@@ -1,13 +1,0 @@
-package ch.epfl.qedit;
-
-public class ResponseLogin extends Response {
-    private User user;
-
-    public ResponseLogin(User user) {
-        this.user = user;
-    }
-
-    public User getUser() {
-        return user;
-    }
-}
