--- app/src/main/java/ch/epfl/qedit/Role.java
@@ -1,8 +1,0 @@
-package ch.epfl.qedit;
-
-public enum Role {
-    ADMIN,
-    EDITOR,
-    CORRECTOR,
-    PARTICIPANT
-}
