--- app/src/main/java/ch/epfl/qedit/User.java
@@ -1,37 +1,0 @@
-package ch.epfl.qedit;
-
-public class User {
-    private String name;
-    private Role role;
-    private String password;
-
-    public User(String name, Role role, String password) {
-        this.name = name;
-        this.role = role;
-        this.password = password;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Role getRole() {
-        return role;
-    }
-
-    public void setRole(Role role) {
-        this.role = role;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
