+++ app/src/main/java/ch/epfl/qedit/frontend/MainActivity.java
@@ -1,7 +1,8 @@
-package ch.epfl.qedit;
+package ch.epfl.qedit.frontend;
 
 import android.os.Bundle;
 import androidx.appcompat.app.AppCompatActivity;
+import ch.epfl.qedit.R;
 
 public class MainActivity extends AppCompatActivity {
